+++ src/test/directives/styleMap_test.ts
@@ -15,91 +15,11 @@
 /// <reference path="../../../node_modules/@types/mocha/index.d.ts" />
 /// <reference path="../../../node_modules/@types/chai/index.d.ts" />
 
-import {classMap, styleMap, ClassInfo, StyleInfo} from '../../directives/css.js';
+import {styleMap, StyleInfo} from '../../directives/styleMap.js';
 import {render} from '../../lib/render.js';
 import {html} from '../../lit-html.js';
 
 const assert = chai.assert;
-
-suite('classMap', () => {
-  let container: HTMLDivElement;
-
-  function renderClassMap(cssInfo: ClassInfo) {
-    render(html`<div class="${classMap(cssInfo)}"></div>`, container);
-  }
-
-  function renderClassMapStatic(cssInfo: ClassInfo) {
-    render(html`<div class="aa ${classMap(cssInfo)}  bb "></div>`, container);
-  }
-
-  setup(() => {
-    container = document.createElement('div');
-  });
-
-  test('adds classes', () => {
-    renderClassMap({ foo: 0, bar: true, zonk: true });
-    const el = container.firstElementChild!;
-    assert.isFalse(el.classList.contains('foo'));
-    assert.isTrue(el.classList.contains('bar'));
-    assert.isTrue(el.classList.contains('zonk'));
-  });
-
-  test('removes classes', () => {
-    renderClassMap({ foo: true, bar: true, baz: true });
-    const el = container.firstElementChild!;
-    assert.isTrue(el.classList.contains('foo'));
-    assert.isTrue(el.classList.contains('bar'));
-    assert.isTrue(el.classList.contains('baz'));
-    renderClassMap({ foo: false, bar: true, baz: false });
-    assert.isFalse(el.classList.contains('foo'));
-    assert.isTrue(el.classList.contains('bar'));
-    assert.isFalse(el.classList.contains('baz'));
-  });
-
-  test('removes omitted classes', () => {
-    renderClassMap({ foo: true, bar: true, baz: true });
-    const el = container.firstElementChild!;
-    assert.isTrue(el.classList.contains('foo'));
-    assert.isTrue(el.classList.contains('bar'));
-    assert.isTrue(el.classList.contains('baz'));
-    renderClassMap({ });
-    assert.isFalse(el.classList.contains('foo'));
-    assert.isFalse(el.classList.contains('bar'));
-    assert.isFalse(el.classList.contains('baz'));
-    assert.equal(el.classList.length, 0);
-  });
-
-  test('works with static classes', () => {
-    renderClassMapStatic({ foo: true});
-    const el = container.firstElementChild!;
-    assert.isTrue(el.classList.contains('aa'));
-    assert.isTrue(el.classList.contains('bb'));
-    assert.isTrue(el.classList.contains('foo'));
-    renderClassMapStatic({});
-    assert.isTrue(el.classList.contains('aa'));
-    assert.isTrue(el.classList.contains('bb'));
-    assert.isFalse(el.classList.contains('foo'));
-  });
-
-  test('throws when used on non-class attribute', () => {
-    assert.throws(() => {
-      render(html`<div id="${classMap({})}"></div>`, container);
-    });
-  });
-
-  test('throws when used in attribute with more than 1 part', () => {
-    assert.throws(() => {
-      render(html`<div class="${'hi'} ${classMap({})}"></div>`, container);
-    });
-  });
-
-  test('throws when used in NodePart', () => {
-    assert.throws(() => {
-      render(html`<div>${classMap({})}</div>`, container);
-    });
-  });
-
-});
 
 suite('styleMap', () => {
   let container: HTMLDivElement;
