+++ test/index.html
@@ -20,7 +20,8 @@
     <script type="module" src="./directives/until_test.js"></script>
     <script type="module" src="./directives/unsafe-html_test.js"></script>
     <script type="module" src="./directives/when_test.js"></script>
-    <script type="module" src="./directives/css_test.js"></script>
+    <script type="module" src="./directives/classMap_test.js"></script>
+    <script type="module" src="./directives/styleMap_test.js"></script>
     <script type="module">
       WCT.loadSuites([
         'shady.html',
